+++ auto/lib/config.lib.ps1
@@ -290,10 +290,11 @@ function Initialize() {
     Set-ConfigValue VersionFile "res\version.txt"
     Set-ConfigValue UserName $null
     Set-ConfigValue UserEmail $null
-    Set-ConfigValue CustomConfigFile "config.ps1"
+    Set-ConfigValue CustomConfigDir "config"
+    Set-ConfigValue CustomConfigFile '$CustomConfigDir$\config.ps1'
     Set-ConfigValue CustomConfigTemplate "res\config.template.ps1"
     Set-ConfigValue AppIndex "res\apps.md"
-    Set-ConfigValue CustomAppIndex "apps.md"
+    Set-ConfigValue CustomAppIndex '$CustomConfigDir$\apps.md'
     Set-ConfigValue CustomAppIndexTemplate "res\apps.template.md"
     Set-ConfigValue DownloadDir "res\download"
     Set-ConfigValue AppResourceBaseDir "res\apps"
@@ -320,6 +321,7 @@ function Initialize() {
     Set-ConfigValue EditorApp "VSCode"
 
     $Script:pathConfigValues = @(
+        "CustomConfigDir",
         "CustomConfigFile",
         "CustomConfigTemplate",
         "AppIndex",